[Graphics] Remove the workaround for older tmux
The workaround was to avoid drawing images when we are performing direct
upload. The problem with the workaround is that it requires full screen
redraw after each direct upload, resulting in poor performance.

Newer versions of tmux shouldn't drop passthrough sequences if the
terminal is too slow to process them, so this workaround shouldn't be
required.