[Graphics] Fix inheritance from the left
We need to be able to inherit missing row/column/4th byte from the cell
to the left even if they have different attributes (like background,
wrap mode, etc, but not foreground obviously). This is somewhat
problematic since xdrawimage is called for cells with exactly the same
attributes. The solution implemented in this commit is to just look at
the cell outside of the strip. Since the cell outside of the strip may
also need to inherit from the left, we just populate the missing data.

Populating missing data is not always correct, e.g. if only the first
column provides the 4th byte, and then we replace the first column, we
may expect the whole image to change, but currently only the first
column will change. This behavior could actually be better in some
cases, and it probably won't matter in practice, so let's keep it for
now.