[Graphics] Don't overwrite classic placements with spaces
When we print space symbols on top of classic placements, don't
overwrite them, just set the background color.  That's in contrast to
Unicode placeholder placements, which must be overwritten.

This is a partial workaround. Ideally, we should have the same behaviour
when erasing part of the line, and even more ideally we need to be able
to have non-space symbols underneath images as well.