[Graphics] Fix diacritic assignment bug
This commit fixes a subtle bug in how properties encoded by diacritics
(row/column/4th id byte) are assigned to cells.

When we see a diacritic, we need to modify the appropriate field of the
last cell. The field is chosen by the number of diacritics that have
already been applied to the cell. Previously, the number of applied
diacritics was determined by checking whether the corresponding fields
were zero. However, this is not correct, because the fields might have
been populated by the procedure that infers missing image cell
properties from the adjacent cell on the left.

This is actually very rare and happens only when there is a time gap
between consecutive diacritics.

The fix is simple: we already remember the number of actually specified
diacritics, so we just use this information instead of checking that
fields are zero.

And also it's worth noting that this inference procedure that actually
fills the missing fields is very convenient, but it's not strictly
correct to do so in some rare cases, which leads to subtle bugs.