[Graphics] Prevent recursive call to twrite
This commit fixes a bug in st that results in a recursive call to twrite
and reprocessing of already processed data coming from the client.

The bug is triggered when the terminal needs to send a lot of data to
the client (like when generating graphics responses), but the client
doesn't expect the data, and the buffer gets full.

To get unstuck, st tries to read and process data coming from the client
by calling ttyread, which calls twrite, and this call is recursive.
This recursive call is not correct, there are many potential issues, and
at least one real issue: since the old data is still in the buf of
ttyread, twrite will reprocess it.

The fix is to avoid the recursive call to twrite. The best we can do is
to read the data into a buffer and process it later. If this buffer gets
full, ttywrite must drop some bytes. I don't see a way around it unless
we want to dynamically grow the size of some buffer.