[Graphics] Fix > 511 columns
There are only 297 diacritics, but a Unicode placeholder image with more
columns may be displayed by inheriting attributes from the left.

However, we have only 9 bits to store the attributes. We can still
display the columns beyond 511, but we cannot remember this information
in the cell. This is problematic for cursor erasure, but it's probably a
rare case.

Also the logic of attribute inheritance is slightly different now:
- We try to reinherit attributes from the left even if we've already
  inherited them and remembered this information on the cell.
- But if there is no information on the left anymore, we may still use
  the information inherited earlier.