[Graphics] Workaround for mac os x freeze on too many responses
Sometimes, when the running app creates lots of images but doesn't read
responses, st may freeze on the write operation. This is a workaround
that makes the file descriptor non-blocking so that we can retry again
and eventually drop the response.

I'm not fully satisfied with this solution, because we may drop
responses partially (admittedly, we already do that). Doing writes
atomically would be preferable, but for that we need to redesign how
ttywrite works.