[Graphics] Optimize pixmap generation
This commit changes how scaled pixmaps are generated.

Previously, we did all the scaling and fitting to the box of cells using
imlib2, which had the following disadvantages:
- the pixmap had to contain transparent pixels in the padding region,
- the pixmap could be much larger than the original image when the box
  of cells was larger.

This commit introduces the following changes:
- Now upscaling is done with XRender during pixmap rendering, so we
  don't need to store pixmaps larger than the image. If we need to
  downscale, we still do it using imlib2.
- Cropping is now also done during pixmap rendering by manipulating
  coordinates and sizes, which eliminates the need to store padding
  pixels.

Note that upscaling with XRender behaves differently from upscaling with
imlib2, which worsens the quality of the rendered images in some cases,
such as when the image is sliced vertically or horizontally. This should
be rare, so it's not very important. However, screenshot tests need to
be updated.